+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComponentFactory.java
@@ -23,9 +23,10 @@ import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.framework.model.SwingEntityEditModel;
 
 import javax.swing.JComponent;
+import java.util.Optional;
 
 /**
- * A factory for {@link ComponentValue} implementations.
+ * A factory for {@link ComponentValue} instances.
  * @param <T> the value type
  * @param <A> the attribute type
  * @param <C> the component type
@@ -33,12 +34,21 @@ import javax.swing.JComponent;
 public interface EntityComponentFactory<T, A extends Attribute<T>, C extends JComponent> {
 
 	/**
-	 * Provides value input components for multiple entity update, override to supply
-	 * specific {@link ComponentValue} implementations for attributes.
-	 * @param attribute the attribute for which to get the ComponentValue
+	 * Provides an input {@link ComponentValue} for editing a single attribute value
+	 * for one or more entities, override to supply a specific {@link ComponentValue}
+	 * implementations for attributes.
+	 * @param attribute the attribute for which to create the {@link ComponentValue}
 	 * @param editModel the edit model used to create foreign key input models
 	 * @param initialValue the initial value to display
-	 * @return a new ComponentValue instance handling input for {@code attribute}
+	 * @return a new {@link ComponentValue} instance handling input for {@code attribute}
 	 */
 	ComponentValue<T, C> componentValue(A attribute, SwingEntityEditModel editModel, T initialValue);
+
+	/**
+	 * Provides a way to override the default attribute caption, when presenting the component to the user.
+	 * @return a caption to use when displaying the component
+	 */
+	default Optional<String> caption() {
+		return Optional.empty();
+	}
 }